chore(join): removed old join page and rename joinpage-new into joinpage
