+++ src/common/Pages.tsx
@@ -18,7 +18,6 @@ import CampaignPreview from 'src/pages/Campaign/preview';
 import Dashboard from 'src/pages/Dashboard';
 import Project from 'src/pages/Dashboard/Project';
 import InsightsPage from 'src/pages/Insights';
-import JoinPage from 'src/pages/JoinPage';
 import LoginPage from 'src/pages/LoginPage';
 import Manual from 'src/pages/Manual';
 import MediaNotFound from 'src/pages/NotFound/MediaNotFound';
@@ -28,7 +27,7 @@ import Profile from 'src/pages/Profile';
 import Template from 'src/pages/Template';
 import Templates from 'src/pages/Templates';
 import Video from 'src/pages/Video';
-import JoinPageNew from 'src/pages/JoinPage-new';
+import JoinPage from 'src/pages/JoinPage';
 import Videos from 'src/pages/Videos';
 import { Redirect } from './Redirect';
 
@@ -74,14 +73,9 @@ const Pages = () => {
                     path={`/${langPrefix}/login`}
                     element={<LoginPage />}
                   />
-                  {/* <Route path={`/${langPrefix}/join`} element={<JoinPage />} />
-                  <Route
-                    path={`/${langPrefix}/join/invites/:profile/:token`}
-                    element={<JoinPage />}
-                  /> */}
                   <Route
                     path={`/${langPrefix}/join/*`}
-                    element={<JoinPageNew />}
+                    element={<JoinPage />}
                   />
 
                   <Route
