+++ src/pages/JoinPage/InvitedUserPage/validationSchema.ts
