--- src/pages/JoinPage/JoinPageLoading.tsx
@@ -1,7 +1,0 @@
-export const JoinPageLoading = () => (
-  <div id="appq-loading-content">
-    <div className="loadingImg">
-      <img alt="loading" src="/static/unguess_loader_single.gif" />
-    </div>
-  </div>
-);
