+++ src/pages/JoinPage/OnboardingPage/Steps/WorkspaceStep.tsx
