+++ src/pages/JoinPage/OnboardingPage/index.tsx
