+++ src/pages/JoinPage/OnboardingPage/validationSchema.ts
