--- src/pages/JoinPage/valuesType.ts
@@ -1,10 +1,0 @@
-export type JoinFormValues = {
-  step: number;
-  email: string;
-  password: string;
-  name: string;
-  surname: string;
-  workspace: string;
-  roleId: number;
-  companySizeId: number;
-};
